Created function that connects to source service and fetch weather data. Fixed the function for fetching city data. When same city was queried with similar names, the weather headline didn't have correct city information, fixed. Created add, list and delete functions for controllers of soucequeryoptions, sourceservice and queryoption models.

+++ src/main/java/com/example/weatherforecastservice/repository/SourceQueryOptionRepository.java
@@ -15,6 +15,6 @@ public interface SourceQueryOptionRepository extends JpaRepository<SourceQueryOp
     SourceQueryOption getCityQueryOption(Long serviceId);
 
     @Query("SELECT q FROM SourceQueryOption q WHERE q.queryOption.Id = ?1 ")
-    List<SourceQueryOption> getSourceQueryOptions(Byte queryOptionId);
+    List<SourceQueryOption> getSourceQueryOptions(Long queryOptionId);
 
 }
